disable Smart Quote/Dash feature on find panel
This new feature substitutes automatically normal quote (") and dash (') symbols with another symbols that are also quote/dash symbol but with an angle, and it is automatically enabled in all NSTextView (and its subclass) if it is build on 10.9.

We know user can disable this feature on the System Preferences system-wide.  However most of average users don't know about it.  Considering the purpose of OgreKit find panel, this feature has less advantages, but just makes user confused.

So I think it's better to disable it on OgreKit find panel.